move input configuration handling to a separate file
